# api/openapi.yaml
openapi: 3.1.0

info:
  title: Fick API
  version: 2026.09.01

paths:
  /api/works/{id}:
    get:
      operationId: getWork

      parameters:
        - name: id
          in: path
          required: true

          schema:
            type: string

      responses:
        "200":
          description: Work

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkResponse"

        "404":
          description: Work not found

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/works/{work_id}/chapters/{chapter_number}:
    get:
      operationId: getChapter

      parameters:
        - name: work_id
          in: path
          required: true
          schema:
            type: string

        - name: chapter_number
          in: path
          required: true
          schema:
            type: string

      responses:
        "200":
          description: Chapter

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterResponse"

  /api/users:
    post:
      operationId: registerUser

      requestBody:
        required: true

        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserRequest"

      responses:
        "201":
          description: User created

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"

        "400":
          description: Registration failed

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "413":
          description: Request body too large

        "415":
          description: Unsupported media type

  /api/users/{id}:
    get:
      operationId: getUser

      parameters:
        - name: id
          in: path
          required: true

          schema:
            type: string

      responses:
        "200":
          description: User

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"

        "404":
          description: User not found

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/users/sessions:
    post:
      operationId: createSession

      requestBody:
        required: true

        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"

      responses:
        "204":
          description: Session created

        "400":
          description: Invalid request

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

        "401":
          description: Invalid credentials

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    WorkVisibility:
      type: string

      enum:
        - public
        - logged_in
        - private

    WorkResponse:
      type: object
      additionalProperties: false

      required:
        - id
        - title
        - summary
        - visibility
        - created_at
        - updated_at
        - chapter_count

      properties:
        id:
          type: string

        title:
          type: string

        summary:
          type: string

        visibility:
          $ref: "#/components/schemas/WorkVisibility"

        created_at:
          type: string
          format: date-time

        updated_at:
          type: string
          format: date-time

        chapter_count:
          type: string

    ChapterResponse:
      type: object
      additionalProperties: false

      required:
        - id
        - number
        - title
        - created_at
        - updated_at
        - content_format
        - content_raw

      properties:
        id:
          type: string

        work_id:
          type: string

        number:
          type: string

        title:
          type: string

        created_at:
          type: string
          format: date-time

        updated_at:
          type: string
          format: date-time

        content_format:
          type: string

        content_raw:
          type: string

    ErrorResponse:
      type: object
      additionalProperties: false

      required:
        - error

      properties:
        error:
          type: string

    RegisterUserRequest:
      type: object
      additionalProperties: false

      required:
        - username
        - password

      properties:
        username:
          type: string
          minLength: 1
          maxLength: 64

        password:
          type: string
          minLength: 15
          maxLength: 256
          writeOnly: true

    UserResponse:
      type: object
      additionalProperties: false

      required:
        - id
        - username
        - created_at
        - updated_at

      properties:
        id:
          type: string

        username:
          type: string

        created_at:
          type: string
          format: date-time

        updated_at:
          type: string
          format: date-time
    
    CreateSessionRequest:
      type: object
      additionalProperties: false

      required:
        - username
        - password

      properties:
        username:
          type: string
          minLength: 1
          maxLength: 64

        password:
          type: string
          minLength: 1
          maxLength: 256
          writeOnly: true