# apps/backend/internal/api/generated/openapi.gen.go
// Package generated provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package generated

import (
	"time"
)

// Defines values for WorkVisibility.
const (
	LoggedIn WorkVisibility = "logged_in"
	Private  WorkVisibility = "private"
	Public   WorkVisibility = "public"
)

// Valid indicates whether the value is a known member of the WorkVisibility enum.
func (e WorkVisibility) Valid() bool {
	switch e {
	case LoggedIn:
		return true
	case Private:
		return true
	case Public:
		return true
	default:
		return false
	}
}

// ChapterResponse defines model for ChapterResponse.
type ChapterResponse struct {
	ContentFormat string    `json:"content_format"`
	ContentRaw    string    `json:"content_raw"`
	CreatedAt     time.Time `json:"created_at"`
	Id            string    `json:"id"`
	Number        string    `json:"number"`
	Title         string    `json:"title"`
	UpdatedAt     time.Time `json:"updated_at"`
	WorkId        *string   `json:"work_id,omitempty"`
}

// CreateSessionRequest defines model for CreateSessionRequest.
type CreateSessionRequest struct {
	Password *string `json:"password,omitempty"`
	Username string  `json:"username"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error string `json:"error"`
}

// RegisterUserRequest defines model for RegisterUserRequest.
type RegisterUserRequest struct {
	Password *string `json:"password,omitempty"`
	Username string  `json:"username"`
}

// UserResponse defines model for UserResponse.
type UserResponse struct {
	CreatedAt time.Time `json:"created_at"`
	Id        string    `json:"id"`
	UpdatedAt time.Time `json:"updated_at"`
	Username  string    `json:"username"`
}

// WorkResponse defines model for WorkResponse.
type WorkResponse struct {
	ChapterCount string         `json:"chapter_count"`
	CreatedAt    time.Time      `json:"created_at"`
	Id           string         `json:"id"`
	Summary      string         `json:"summary"`
	Title        string         `json:"title"`
	UpdatedAt    time.Time      `json:"updated_at"`
	Visibility   WorkVisibility `json:"visibility"`
}

// WorkVisibility defines model for WorkVisibility.
type WorkVisibility string

// RegisterUserJSONRequestBody defines body for RegisterUser for application/json ContentType.
type RegisterUserJSONRequestBody = RegisterUserRequest

// CreateSessionJSONRequestBody defines body for CreateSession for application/json ContentType.
type CreateSessionJSONRequestBody = CreateSessionRequest